Changed user ID from 16 bit int to 32 bit in. Just means it's future-proofed for ~2 billion users.

+++ user/init_user.go
@@ -7,9 +7,9 @@ import (
 
 //This initializes a user upon log-in. It returns the userID given, for later usage.
 //TODO: Basically all of the extra functionality in this
-func InitUser(userID []byte) uint16 {
-	var user uint16
-	user = binary.BigEndian.Uint16(userID) //Thanks to https://stackoverflow.com/questions/11184336/how-to-convert-from-byte-to-int-in-go-programming
+func InitUser(userID []byte) uint32 {
+	var user uint32
+	user = binary.BigEndian.Uint32(userID) //Thanks to https://stackoverflow.com/questions/11184336/how-to-convert-from-byte-to-int-in-go-programming
 	return user
 }
